requirements.txt
python-gitlab >= 3.0.0
pandas
configparser
openpyxl
